feat: Refactore code
